refactor: add history utils
