+++ src/components/history/History.tsx
@@ -7,12 +7,7 @@ import { RoutePaths } from '@constants/routePaths';
 import { UILink } from '@ui/UILink';
 import { UIButton } from '@ui/UIButton';
 import { useEffect, useState } from 'react';
-
-interface IHistoryRequest {
-  metod: string;
-  url: string;
-  requestDate: string;
-}
+import { addHistoryData, IHistoryRequest } from '@utils/history';
 
 function History() {
   const t = useTranslations('WelcomePage');
@@ -22,19 +17,6 @@ function History() {
     const historyFromLS = localStorage.getItem('history-requests');
     if (historyFromLS != null && historyFromLS.length) setHistory(JSON.parse(historyFromLS));
   }, []);
-
-  const addHistoryData = (data: Record<string, string>) => {
-    const key = 'history-requests';
-    const currentData = localStorage.getItem(key) || '';
-    const requestDate = Date.now();
-    const newData =
-      currentData.length ? [...JSON.parse(currentData), { ...data, requestDate }] : [{ ...data, requestDate }];
-    localStorage.setItem(key, JSON.stringify(newData));
-  };
-
-  const test = () => {
-    console.log(history, history.length);
-  };
 
   const generateHistoryItem = () => {
     addHistoryData({
@@ -54,11 +36,8 @@ function History() {
       <div className="flex flex-row mt-4 mb-4 gap-[10px] justify-center">
         <UIButton onClick={generateHistoryItem}>Genetare request</UIButton>
         <UIButton onClick={clearHistory}>Clear history</UIButton>
-        <UIButton onClick={test}>Log history state</UIButton>
       </div>
-
       <h2 className="text-4xl mt-2 mb-2">{'History Requests'}</h2>
-
       {history.length ?
         <div className="flex flex-col gap-[10px] mb-[20px]">
           {history
@@ -82,7 +61,6 @@ function History() {
             })}
         </div>
       : <>
-          {' '}
           <p>{`You haven't executed any requests. It's empty here. Try:`}</p>
           <UILink
             href={RoutePaths.REST}
